# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: Donald O. Oduol
    url: https://github.com/oduold
  description: Web inventory API for a supermarket store
  title: Brandlogs Supermarket API
  version: 0.0.1
servers:
- description: Development Server
  url: http://localhost:8080
...
paths:
...
  /locations:
    get:
      description: List of Locations
      operationId: locations
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Location'
                type: array
          description: A paged array of Locations
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Retrieve all Locations
      tags:
      - Locations
      x-accepts: application/json
      x-tags:
      - tag: Locations
    post:
      description: Add a Location
      operationId: addLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
          description: A Location's data
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Add a Location
      tags:
      - Locations
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Locations
  /locations/{id}:
    delete:
      description: Location data by id
      operationId: locationDelete
      parameters:
      - description: resource id
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: A Location's deleted
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Delete location data by id
      tags:
      - Location
      x-accepts: application/json
      x-tags:
      - tag: Location
    get:
      description: Location data by id
      operationId: location
      parameters:
      - description: resource id
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
          description: A Location's data
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Retrieve location data by id
      tags:
      - Location
      x-accepts: application/json
      x-tags:
      - tag: Location
    put:
      description: Update Location data by id
      operationId: locationEdit
      parameters:
      - description: resource id
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
        description: Update location properties to be changed
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
          description: A Location's data
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Update location data by id
      tags:
      - Location
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Location
...
components:
  schemas:
...
    Location:
      example:
        code: code
        name: name
        id: 0
      properties:
        id:
          format: int64
          type: integer
        code:
          type: string
        name:
          type: string
      type: object
...